webapp/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
djangorestframework-gis==1.0
geopandas==0.14.4
python-decouple==3.8
django-cors-headers==4.2.0
psycopg2-binary==2.9.9
django-filter==24.2
djangorestframework-simplejwt==5.3.1
